requirements.txt
lxml==4.0.0
Appium-Python-Client==0.26
selenium
